# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentación de la API de CentralizadorBackEnd",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo de Desarrollo",
      "email": "devteam@example.com"
    },
    "description": "API REST para la gestión de datos de la plataforma CogniCare, incluyendo estudiantes, entrenadores y entrenamientos cognitivos.\n\nAutenticación\n\n **IMPORTANTE:** La mayoría de los endpoints requieren un token de autenticación (JWT Bearer).\n\nPara acceder a los endpoints protegidos, sigue estos pasos:\n1. **Obtén tu token:** Realiza una petición `POST` a `/api/login/admin` o `/api/login/entrenador` con tus credenciales.\n2. **Autoriza tus peticiones:** Haz clic en el botón **\"Authorize\"** (arriba a la derecha) y pega tu token en el campo `bearerAuth` con el formato `Bearer <TU_TOKEN_JWT>`.\n\nRoles y Permisos\n\nLa API cuenta con dos roles principales con diferentes niveles de acceso:\n\n### **Admin**\n- Gestión completa de entrenadores (CRUD).\n- Visualización de todos los informes de entrenamiento.\n- Acceso total a las utilidades del sistema.\n\n### **Entrenador**\n- Gestión de estudiantes y sus entrenamientos cognitivos.\n- Creación y seguimiento de sesiones de entrenamiento.\n- Visualización de informes de los estudiantes asignados."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido al hacer login. Formato: `Bearer <tu_token>`"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensaje de error."
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "description": "Mensaje de validación del error."
                },
                "param": {
                  "type": "string",
                  "description": "Parámetro asociado al error."
                },
                "location": {
                  "type": "string",
                  "description": "Ubicación del error (e.g., \"body\")."
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/entrenadores/facultad-entrenador/{correo}": {
      "get": {
        "summary": "Obtiene las facultades asociadas a un entrenador por su correo electrónico.",
        "tags": [
          "Entrenadores"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "correo",
            "required": true,
            "description": "Correo electrónico del entrenador.",
            "schema": {
              "type": "string",
              "format": "email",
              "example": "entrenador@example.com"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de facultades asociadas al entrenador.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "facultad_id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID de la facultad.",
                        "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
                      },
                      "nombre_facultad": {
                        "type": "string",
                        "description": "Nombre de la facultad.",
                        "example": "Facultad de Ingeniería"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Correo electrónico inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Correo electrónico inválido."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron facultades para el entrenador con el correo proporcionado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "No se encontraron facultades para el entrenador con el correo proporcionado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor al obtener facultades."
                    },
                    "error": {
                      "type": "string",
                      "example": "Detalle del error."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/entrenadores/datosporcorreo/{correo}": {
      "get": {
        "summary": "Obtiene los datos de un entrenador por su correo electrónico.",
        "tags": [
          "Entrenadores"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "correo",
            "required": true,
            "description": "Correo electrónico del entrenador.",
            "schema": {
              "type": "string",
              "format": "email",
              "example": "monica.melendez@uco.edu.co"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del entrenador obtenidos exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Datos del entrenador obtenidos exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "numeroDocumento": {
                          "type": "string",
                          "description": "Número de documento del entrenador.",
                          "example": "123456789"
                        },
                        "siglaTipoDocumento": {
                          "type": "string",
                          "description": "Sigla del tipo de documento.",
                          "example": "CC"
                        },
                        "nombres": {
                          "type": "string",
                          "description": "Nombres del entrenador.",
                          "example": "Juan Carlos"
                        },
                        "apellidos": {
                          "type": "string",
                          "description": "Apellidos del entrenador.",
                          "example": "Pérez García"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "El correo electrónico debe tener un formato válido."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontró un entrenador con el correo proporcionado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "No se encontró un entrenador con el correo electrónico proporcionado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/entrenadores/informes": {
      "get": {
        "summary": "Obtiene la lista de entrenamientos asignados al entrenador autenticado.",
        "tags": [
          "InformesEntrenador"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de entrenamientos obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Lista de entrenamientos asignados obtenida exitosamente."
                    },
                    "total": {
                      "type": "integer",
                      "example": 1
                    },
                    "entrenamientos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "entrenamientoid": {
                            "type": "string",
                            "format": "uuid",
                            "description": "ID del entrenamiento cognitivo."
                          },
                          "estudiantenombres": {
                            "type": "string",
                            "description": "Nombres del estudiante.",
                            "example": "Pedro León"
                          },
                          "estudianteapellidos": {
                            "type": "string",
                            "description": "Apellidos del estudiante.",
                            "example": "García Pérez"
                          },
                          "estudiantetipodocumento": {
                            "type": "string",
                            "description": "Sigla del tipo de documento del estudiante.",
                            "example": "CC"
                          },
                          "estudiantenumerodocumento": {
                            "type": "string",
                            "description": "Número de documento del estudiante.",
                            "example": "1036402111"
                          },
                          "estadoentrenamiento": {
                            "type": "string",
                            "description": "Estado actual del entrenamiento.",
                            "example": "Por Iniciar"
                          },
                          "fechainicioentrenamiento": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Fecha de inicio del entrenamiento."
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. El token no fue proporcionado o es inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Acceso prohibido. El usuario no tiene el rol de 'Entrenador'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}